feat database connect

+++ frontend/src/api/classroom.ts
@@ -1,5 +1,23 @@
 import api from "./client";
 
+const createClassroom = async (name: string, description: string = "") => {
+    const response = await api.post("/classroom/new", 
+        {
+            name,
+            description,
+        },
+        { withCredentials: true }
+    );
+    return response.data;
+}
+
+const deleteClassroom = async (classroomId: number) => {
+    const response = await api.delete(`/classroom/delete/${classroomId}`, {
+        withCredentials: true,
+    });
+    return response.data;
+}
+
 const getAllClassrooms = async () => {
     const response = await api.get("/classroom/", { withCredentials: true });
     return response.data;
@@ -11,11 +29,10 @@ const getAllStudentsInClassroom = async (classroomId: number) => {
 }
 
 const removeStudentFromClassroom = async (classroomId: number, userId: number) => {
-    const response = await api.delete(`/classroom/${classroomId}/`, {
-        data: { userId },
+    const response = await api.delete(`/classroom/${classroomId}/${userId}`, {
         withCredentials: true,
     });
     return response.data;
 }
 
-export { getAllClassrooms, getAllStudentsInClassroom, removeStudentFromClassroom }; 
+export { createClassroom, deleteClassroom, getAllClassrooms, getAllStudentsInClassroom, removeStudentFromClassroom };
